# api/oas/server/server.go
// Code generated by woco, DO NOT EDIT.

package server

import (
	"errors"
	"net/http"

	"github.com/gin-gonic/gin"
	"github.com/tsingsun/woocoo/web/handler"
	"github.com/woocoos/msgcenter/api/oas"
)

// RegisterGeneralHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterGeneralHandlers(router *gin.RouterGroup, si oas.GeneralService) {
	router.GET("/status", wrapGetStatus(si))
}

// RegisterReceiverHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterReceiverHandlers(router *gin.RouterGroup, si oas.ReceiverService) {
	router.GET("/receivers", wrapGetReceivers(si))
}

// RegisterSilenceHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterSilenceHandlers(router *gin.RouterGroup, si oas.SilenceService) {
	router.DELETE("/silence/:silenceID", wrapDeleteSilence(si))
	router.GET("/silence/:silenceID", wrapGetSilence(si))
	router.GET("/silences", wrapGetSilences(si))
	router.POST("/silences", wrapPostSilences(si))
}

// RegisterAlertHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterAlertHandlers(router *gin.RouterGroup, si oas.AlertService) {
	router.GET("/alerts", wrapGetAlerts(si))
	router.POST("/alerts", wrapPostAlerts(si))
}

// RegisterPushHandlers creates http.Handler with routing matching OpenAPI spec.
func RegisterPushHandlers(router *gin.RouterGroup, si oas.PushService) {
	router.POST("/push", wrapPostPush(si))
}

func wrapGetStatus(si oas.GeneralService) func(c *gin.Context) {
	return func(c *gin.Context) {
		resp, err := si.GetStatus(c)
		if err != nil {
			c.Error(err)
			return
		}
		handler.NegotiateResponse(c, http.StatusOK, resp, []string{"application/json"})
	}
}

func wrapGetReceivers(si oas.ReceiverService) func(c *gin.Context) {
	return func(c *gin.Context) {
		resp, err := si.GetReceivers(c)
		if err != nil {
			c.Error(err)
			return
		}
		handler.NegotiateResponse(c, http.StatusOK, resp, []string{"application/json"})
	}
}

func wrapDeleteSilence(si oas.SilenceService) func(c *gin.Context) {
	return func(c *gin.Context) {
		var req oas.DeleteSilenceRequest
		if err := c.ShouldBindUri(&req.UriParams); err != nil {
			handler.AbortWithError(c, http.StatusBadRequest, err)
			return
		}
		err := si.DeleteSilence(c, &req)
		if err != nil {
			c.Error(err)
			return
		}
	}
}

func wrapGetSilence(si oas.SilenceService) func(c *gin.Context) {
	return func(c *gin.Context) {
		var req oas.GetSilenceRequest
		if err := c.ShouldBindUri(&req.UriParams); err != nil {
			handler.AbortWithError(c, http.StatusBadRequest, err)
			return
		}
		resp, err := si.GetSilence(c, &req)
		if err != nil {
			c.Error(err)
			return
		}
		if resp == nil {
			handler.AbortWithError(c, http.StatusBadRequest, errors.New("A silence with the specified ID was not found"))
			return
		}
		handler.NegotiateResponse(c, http.StatusOK, resp, []string{"application/json"})
	}
}

func wrapGetSilences(si oas.SilenceService) func(c *gin.Context) {
	return func(c *gin.Context) {
		var req oas.GetSilencesRequest
		if err := c.ShouldBind(&req); err != nil {
			handler.AbortWithError(c, http.StatusBadRequest, err)
			return
		}
		resp, err := si.GetSilences(c, &req)
		if err != nil {
			c.Error(err)
			return
		}
		handler.NegotiateResponse(c, http.StatusOK, resp, []string{"application/json"})
	}
}

func wrapPostSilences(si oas.SilenceService) func(c *gin.Context) {
	return func(c *gin.Context) {
		var req oas.PostSilencesRequest
		if err := c.ShouldBind(&req.PostableSilence); err != nil {
			handler.AbortWithError(c, http.StatusBadRequest, err)
			return
		}
		resp, err := si.PostSilences(c, &req)
		if err != nil {
			c.Error(err)
			return
		}
		if resp == nil {
			handler.AbortWithError(c, http.StatusBadRequest, errors.New("A silence with the specified ID was not found"))
			return
		}
		handler.NegotiateResponse(c, http.StatusOK, resp, []string{"application/json"})
	}
}

func wrapGetAlerts(si oas.AlertService) func(c *gin.Context) {
	return func(c *gin.Context) {
		var req oas.GetAlertsRequest
		if err := c.ShouldBind(&req.Body); err != nil {
			handler.AbortWithError(c, http.StatusBadRequest, err)
			return
		}
		resp, err := si.GetAlerts(c, &req)
		if err != nil {
			c.Error(err)
			return
		}
		handler.NegotiateResponse(c, http.StatusOK, resp, []string{"application/json"})
	}
}

func wrapPostAlerts(si oas.AlertService) func(c *gin.Context) {
	return func(c *gin.Context) {
		var req oas.PostAlertsRequest
		if err := c.ShouldBind(&req.PostableAlerts); err != nil {
			handler.AbortWithError(c, http.StatusBadRequest, err)
			return
		}
		err := si.PostAlerts(c, &req)
		if err != nil {
			c.Error(err)
			return
		}
	}
}

func wrapPostPush(si oas.PushService) func(c *gin.Context) {
	return func(c *gin.Context) {
		var req oas.PostPushRequest
		if err := c.ShouldBind(&req.PushData); err != nil {
			handler.AbortWithError(c, http.StatusBadRequest, err)
			return
		}
		err := si.PostPush(c, &req)
		if err != nil {
			c.Error(err)
			return
		}
	}
}